ruoyi 管理端 完善

+++ ruoyi-admin/src/main/java/com/ruoyi/web/controller/ai_interview/BannerController.java
@@ -8,6 +8,7 @@ import com.ruoyi.common.core.domain.AjaxResult;
 import com.ruoyi.common.core.page.TableDataInfo;
 import com.ruoyi.common.enums.BusinessType;
 import com.ruoyi.common.utils.MinioUtil;
+import org.apache.logging.log4j.util.Strings;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
@@ -53,13 +54,19 @@ public class BannerController extends BaseController {
         if (bannerService.checkTitleExsit(banner)) {
             return error("新增banner'" + banner.getTitle() + "'失败，名称已存在");
         }
-        // 处理上传的文件
-        String originalFilename = file.getOriginalFilename();
-        assert originalFilename != null;
-        String fileExtension = originalFilename.substring(originalFilename.lastIndexOf("."));
-        String newFilename = sdf.format(System.currentTimeMillis()) + fileExtension;
-        String uploadFileUrl = minioUtil.uploadFile(file, newFilename);
+        String uploadFileUrl = "";
+        if(Strings.isEmpty(banner.getUrl())){
+            // 处理上传的文件
+            String originalFilename = file.getOriginalFilename();
+            assert originalFilename != null;
+            String fileExtension = originalFilename.substring(originalFilename.lastIndexOf("."));
+            String newFilename = sdf.format(System.currentTimeMillis()) + fileExtension;
+            uploadFileUrl = minioUtil.uploadFile(file, newFilename);
+        }else {
+            uploadFileUrl = banner.getUrl();
+        }
         banner.setUrl(uploadFileUrl);
+        banner.setImage(uploadFileUrl);
         return toAjax(bannerService.save(banner));
 
     }
@@ -79,7 +86,8 @@ public class BannerController extends BaseController {
         if (bannerService.checkTitleExsit(banner)) {
             return error("编辑banner'" + banner.getTitle() + "'失败，名称已存在");
         }
-        if(file != null){
+        String uploadFileUrl = banner.getUrl();
+        if(Strings.isEmpty(uploadFileUrl) && file != null){
             // 删除旧的文件
             minioUtil.deleteFile(minioUtil.extractFileNameFromUrl(banner.getUrl()));
             // 处理上传的文件
@@ -87,9 +95,11 @@ public class BannerController extends BaseController {
             assert originalFilename != null;
             String fileExtension = originalFilename.substring(originalFilename.lastIndexOf("."));
             String newFilename = sdf.format(System.currentTimeMillis()) + fileExtension;
-            String uploadFileUrl = minioUtil.uploadFile(file, newFilename);
-            banner.setUrl(uploadFileUrl);
+            uploadFileUrl = minioUtil.uploadFile(file, newFilename);
+
         }
+        banner.setUrl(uploadFileUrl);
+        banner.setImage(uploadFileUrl);
         return toAjax(bannerService.updateById(banner));
     }
